requirements.txt
ipykernel
langchain
langchain-core
langchain_community
langchain-openai
langchain-chroma
openai
langchain-anthropic
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub
python-dotenv
scikit-learn
pandas
pypdf
bs4
arxiv
pymupdf
nltk
wikipedia
streamlit
langchain-text-splitters
chromadb
faiss-cpu
langchain-groq
grandalf
langchain_experimental
weaviate-client
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
unstructured[pdf]
unstructured[pptx]
pytube
numexpr
huggingface-hub
pymongo
sentence_transformers
datasets
requests
pinecone 
poppler-utils
unstructured-pytesseract
langchain-astradb